Support dtype=np.float32 even when jax_enable_x64 is set True
To enable double precision float in JAX, use:

    from jax.config import config
    config.update("jax_enable_x64", True)

+++ mod/pray/icond.py
@@ -33,4 +33,4 @@ def cam(rij, ab):
     return np.array([
         [0, x, y, z],
         [1, si * cj, si * sj, ci],
-    ])
+    ], dtype=ab.dtype)
